- created a song demo scene with animations - started new motion controller state script
