--- project/Assets/EyeBlinkControllerProfile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 47037bbaf91f39348a2b37ec678b7f97
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
