+++ project/Assets/Sounds/music.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4b567da2624b75448bf061fc9853b0e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
