+++ project/Assets/Sprites/characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4f84c4a4b473a4541ad0c8906f699cab
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
