+++ project/Assets/animations/characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ed9d10f167834c245bed5c860bd5bbbc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
