+++ project/Assets/animations/characters/ghost-king-body-wrapper.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6dd2a07ed20b70140aef0ce0f83ea669
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
