+++ project/Assets/animations/characters/ghost-king-demo.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 28b9e7c1c47dfd243a452fcca3c4848e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
