+++ project/Assets/animations/characters/ghost-king-song-sequence.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fec2f6e2483b3e94d8cf76cd115e14c5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
