+++ project/Assets/animations/characters/ghost-king-wrapper.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1789f0a55779c2c4dacbad2d9b5d2018
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
