+++ project/Assets/animations/hero/sit-listen-music.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 473ababb4d418b94f86c794fd782a3a6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
