+++ project/Assets/scripts/HeroMotionController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 32e5ce38ce36f5443b5a96245f3c753b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
