+++ project/Library/LastSceneManagerSetup.txt
@@ -1,5 +1,5 @@
 sceneSetups:
-- path: Assets/scenes/puppet_staging.unity
+- path: Assets/scenes/BrokenPool.unity
   isLoaded: 1
   isActive: 1
   isSubScene: 0
