+++ project/Logs/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,3 +1,4 @@
 Base path: 'C:/Program Files/Unity/Hub/Editor/2020.3.25f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2020.3.25f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
 
+Cmd: shutdown
